Extract out Geant4ResultsFileParser
